LayoutTests/fast/frames/seamless/seamless-hyperlink-expected.txt

FAIL iframe.contentDocument.body.textContent should be PASS. Was FAIL                window.onload=function() {            frames[0].document.getElementsByTagName('a')[0].click();        }    .
